Added user_id to comment class and table in DB; Added authentication for put, post and delete operations.

+++ src/main/java/com/example/forummanagementsystem/controllers/CommentController.java
@@ -1,6 +1,9 @@
 package com.example.forummanagementsystem.controllers;
 
-
+import com.example.forummanagementsystem.exceptions.AuthorizationException;
+import com.example.forummanagementsystem.helpers.AuthenticationHelper;
+import com.example.forummanagementsystem.models.Post;
+import com.example.forummanagementsystem.models.User;
 import com.example.forummanagementsystem.models.Comment;
 import com.example.forummanagementsystem.models.CommentDto;
 
@@ -10,6 +13,7 @@ import com.example.forummanagementsystem.services.mappers.CommentMapper;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -24,7 +28,6 @@ import com.example.forummanagementsystem.exceptions.EntityNotFoundException;
 import org.springframework.http.HttpStatus;
 
 
-
 @RestController
 @RequestMapping("api/comments")
 
@@ -35,11 +38,14 @@ public class CommentController {
 
     private final CommentMapper modelMapper;
 
+    private final AuthenticationHelper authenticationHelper;
+
     @Autowired
-    public CommentController(CommentService commentService, CommentMapper modelMapper) {
+    public CommentController(CommentService commentService, CommentMapper modelMapper, AuthenticationHelper authenticationHelper) {
         this.commentService = commentService;
         this.modelMapper = modelMapper;
 
+        this.authenticationHelper = authenticationHelper;
     }
 
     @GetMapping
@@ -57,9 +63,10 @@ public class CommentController {
     }
 
     @PostMapping
-    public Comment create(@Valid @RequestBody CommentDto commentDto) {
+    public Comment create(@RequestHeader HttpHeaders headers, @Valid @RequestBody CommentDto commentDto) {
         try {
-            Comment comment = modelMapper.fromDtoComment(commentDto);
+            User user = authenticationHelper.tryGetUser(headers);
+            Comment comment = modelMapper.dtoToObjectComment(commentDto, user);
             commentService.create(comment);
             return comment;
         } catch (EntityDuplicateException e) {
@@ -68,24 +75,35 @@ public class CommentController {
     }
 
     @PutMapping("/{id}")
-    public Comment update(@PathVariable int id, @Valid @RequestBody CommentDto commentDto) {
+    public Comment update(@RequestHeader HttpHeaders headers, @PathVariable int id, @Valid @RequestBody CommentDto commentDto) {
         try {
-            Comment comment = modelMapper.fromDtoComment(commentDto,id);
-            commentService.update(comment);
-            return comment;
+            User user = authenticationHelper.tryGetUser(headers);
+            Comment commentToUpdate = commentService.getById(id);
+            Comment newComment = modelMapper.dtoToObjectComment(commentDto, commentToUpdate.getUser());
+            authenticationHelper.checkPermissions(commentService.getById(id).getUser().getId(), user);
+            newComment.setCommentId(id);
+            commentService.update(newComment, commentToUpdate.getUser());
+            return newComment;
+
         } catch (EntityNotFoundException e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
         } catch (EntityDuplicateException e) {
             throw new ResponseStatusException(HttpStatus.CONFLICT, e.getMessage());
+        } catch (AuthorizationException e) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
         }
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id) {
+    public void delete(@RequestHeader HttpHeaders headers, @PathVariable int id) {
         try {
+            User user = authenticationHelper.tryGetUser(headers);
+            authenticationHelper.checkPermissions(commentService.getById(id).getUser().getId(), user);
             commentService.delete(id);
         } catch (EntityNotFoundException e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
+        } catch (AuthorizationException e) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage());
         }
     }
 }
